add date input to compiler
