--- ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/Block/formPrimitives/FormStringInput/TextInput.tsx
@@ -1,12 +1,0 @@
-import { FormStringInputType } from "../../../../schema/blocks/form/stringInput";
-import Input from "~/design/Input";
-import { InputProps } from "./types";
-
-type TextInputProps = InputProps & {
-  variant: FormStringInputType["variant"];
-};
-
-export const TextInput = (props: TextInputProps) => {
-  const { id, defaultValue, variant } = props;
-  return <Input id={id} defaultValue={defaultValue} type={variant} />;
-};
