--- ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/Block/formPrimitives/FormStringInput/types.ts
@@ -1,4 +1,0 @@
-export type InputProps = {
-  defaultValue: string;
-  id: string;
-};
